fastAPI/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.109.2
fonttools==4.53.0
graphql-core==3.2.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mlxtend==0.23.1
mypy-extensions==1.0.0
numpy==1.26.0
openpyxl==3.1.3
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
piccolo==1.7.0
piccolo_admin==1.4.0
piccolo_api==1.4.0
pillow==10.3.0
platformdirs==4.2.2
priority==2.0.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
scikit-learn==1.5.0
scikit-surprise==1.1.4
scipy==1.13.1
seaborn==0.13.2
setuptools==70.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
strawberry-graphql==0.234.0
surprise==0.1
targ==0.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.27.1
wordcloud==1.9.3
wsproto==1.2.0
lightfm==1.17